weapon damage is now an int, on client shoot it passes weapon damage to cmd player shoot function which now sets an instance of a player based off game manager get player id of string argument then calls yet to be made take damage function argument damage from player

+++ spaceMultiplayer/Assets/Scripts/PlayerWeapon.cs
@@ -5,6 +5,6 @@ public class PlayerWeapon
 {
     public string name = "Glock";
 
-    public float damage = 20f;
+    public int damage = 20;
     public float range = 100f;
 }
